refactor: moving the datafactory client
